shell/events: handle common events in trx
